add more for correct_mw.cpp

+++ src/RawMessageConnection.cpp
@@ -42,7 +42,7 @@ void RawMessageConnection::sendRawMessage(RawMessage *m,
 
 std::ostream &operator<<(std::ostream &os, const RawMessage &msg)
 {
-    os << "{RawMessage Type: " << msg.type << ", node_id: " << msg.node_id
+    os << "{RawMessage Type: " << (int) msg.type << ", node_id: " << msg.node_id
        << ", app_id: " << msg.app_id << ", addr: " << msg.addr
        << ", level: " << msg.level
        << ", inlined_buffer[0]: " << msg.inlined_buffer[0] << "}";
